build fix

+++ contracts/examples/adder/interact/src/basic_interact_cli.rs
@@ -30,7 +30,7 @@ pub enum InteractCliCommand {
 pub struct AddArgs {
     /// The value to add
     #[arg(short = 'v', long = "value", verbatim_doc_comment)]
-    pub value: u64,
+    pub value: u32,
 }
 
 #[derive(Default, Clone, PartialEq, Eq, Debug, Args)]
